refactor out api code
