+++ server/cron/cron.js
@@ -14,7 +14,7 @@ const { fetchStaleStocksFromDb, fetchAndUpdateStockData } = require('./helpers')
 // const cronSchedule = '0 9,12 * * 1-5';
 
 // run every 15 minutes instead but do it on a stale basis
-const CRON_SCHEDULE = '*/15 * * * *';
+const CRON_SCHEDULE = '*/15 * * * * *';
 
 const cronJob = cron.schedule(
   CRON_SCHEDULE,
